2016-10-28 Roman - !!:FIX:DB.LibsDB.OscadaLibs: DevLib.IEC60870: Fix for the new behavior of attributes of Object free by EVAL set.     !!:Self implementing for EMail sent by SMTP is added as an user protocol's function, implemented: pure, SSL/TLS, AUTH.     !!:Notifications dispatcher Controllers.ntfDispatch expanded by using self SMTP implementing for EMail send. - !!!!:SYS: More functions set to constant after some objects appending by a copy constructor.     !!!!:More error and warning fixes from PVS-Studio report, thanks to Maksim Kochetkov:         TCntrNode: Possible wrong into copy node function by unsigned group's index.         TCntrNode,TConfig,TRegExp,Archive,DB.{MySQL,PostgreSQL},DAQ.{AMRDevs,JavaLikeCalc},Special.FLibSYS,UI.{Vision,WebVision}:             Possible collision into a function argument or an internal variable name and the class method.         TVariant: Middle strings reallocation is set to the result check.         XMLNodeObj,DAQ.{*,AMRDevs,Siemens},Archive.{*,FSArch},UI.QTCfg: Some logic made simply.     !!: Adapting and some fixing for build by GCC v6 with C++11 by default. - DAQ.{Templates,JavaLikeCalc},Protocol.ModBus,UI.VCAEngine: Items deletion procedure termination at DB error is added. - FIX:Special.FLibSYS: FFT buffers "free" fix, instead "delete". - !:Transport.SSL: Expanded by methods TLSv1.1, TLSv1.2, DTLSv1. - UI.QTStarter: Adapted to Qt5 specific method QTranslator::translate(). - FIX:UI.VCAEngine: Virtual function PrWidget::setEnable() fixed for second argument.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2430 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tmess.cpp
@@ -139,7 +139,7 @@ void TMess::putArg( const char *categ, int8_t level, const char *fmt, va_list ap
 
     if(mLogDir & DIR_SYSLOG) {
 	int level_sys;
-	switch(abs(level)) {
+	switch((int8_t)abs(level)) {
 	    case Debug:		level_sys = LOG_DEBUG;	break;
 	    case Info:		level_sys = LOG_INFO;	break;
 	    case Notice:	level_sys = LOG_NOTICE;	break;
@@ -235,14 +235,14 @@ string TMess::translGet( const string &base, const string &lang, const string &s
 		    req.elem().fldAdd(new TFld(tStrVl.c_str(),"Tr",TFld::String,0));
 		    req.cfg("base").setS(base);
 		}
-		if(SYS->db().at().dataGet(*iA+"."mess_TrUApiTbl,"/"mess_TrUApiTbl,req,false,true)) {
+		if(SYS->db().at().dataGet(*iA+"." mess_TrUApiTbl,"/" mess_TrUApiTbl,req,false,true)) {
 		    rez = req.cfg(tStrVl).getS();
 		    break;
 		}
 		// Create new record into the translation table of the data source
 		else if((iA+1) == addrs.rend() && lang2CodeBase().size()) {
 		    if(lang2CodeBase() == trLang) req.elem().fldDel(req.elem().fldId(tStrVl.c_str()));
-		    SYS->db().at().dataSet(*iA+"."mess_TrUApiTbl, "/"mess_TrUApiTbl, req, false, true);
+		    SYS->db().at().dataSet(*iA+"." mess_TrUApiTbl, "/" mess_TrUApiTbl, req, false, true);
 		}
 	    }
 	    trMessCache[cKey] = CacheEl(rez, SYS->sysTm());
@@ -363,11 +363,11 @@ void TMess::translReg( const string &mess, const string &src, const string &prms
 	MtxAlloc res(mRes, true);
 	for(unsigned i_l = 0; i_l < ls.size(); i_l++)
 	    if(ls[i_l] == DB_CFG)
-		for(int io_cnt = 0; SYS->db().at().dataSeek("","/"mess_TrUApiTbl,io_cnt++,req,false,&full); )
-		    trMessIdx[req.cfg("base").getS()]["cfg:/"mess_TrUApiTbl] = prms;
+		for(int io_cnt = 0; SYS->db().at().dataSeek("","/" mess_TrUApiTbl,io_cnt++,req,false,&full); )
+		    trMessIdx[req.cfg("base").getS()]["cfg:/" mess_TrUApiTbl] = prms;
 	    else
-		for(int io_cnt = 0; SYS->db().at().dataSeek(ls[i_l]+"."mess_TrUApiTbl,"",io_cnt++,req,false,&full); )
-		    trMessIdx[req.cfg("base").getS()]["db:"+ls[i_l]+"."mess_TrUApiTbl"#base"] = prms;
+		for(int io_cnt = 0; SYS->db().at().dataSeek(ls[i_l]+"." mess_TrUApiTbl,"",io_cnt++,req,false,&full); )
+		    trMessIdx[req.cfg("base").getS()]["db:"+ls[i_l]+"." mess_TrUApiTbl "#base"] = prms;
     }
     else {
 	if(TSYS::strNoSpace(mess).empty()) return;
